+++ lib/eager_eye/configuration.rb
@@ -8,7 +8,7 @@ module EagerEye
     DEFAULT_DETECTORS = %i[
       loop_association serializer_nesting missing_counter_cache
       custom_method_query count_in_iteration callback_query
-      pluck_to_array delegation_n_plus_one
+      pluck_to_array delegation_n_plus_one decorator_n_plus_one
     ].freeze
 
     DEFAULT_SEVERITY_LEVELS = {
@@ -19,7 +19,8 @@ module EagerEye
       count_in_iteration: :warning,
       callback_query: :warning,
       pluck_to_array: :warning,
-      delegation_n_plus_one: :warning
+      delegation_n_plus_one: :warning,
+      decorator_n_plus_one: :warning
     }.freeze
 
     VALID_SEVERITIES = %i[info warning error].freeze